+++ loglan_core/__init__.py
@@ -8,29 +8,25 @@ __author__ = "torrua"
 __copyright__ = "Copyright 2024, loglan_core project"
 __email__ = "torrua@gmail.com"
 
-from loglan_core.connect_tables import (
+from loglan_core.service.connect_tables import (
     t_connect_authors,
     t_connect_words,
     t_connect_keys,
 )
-
-from .base import BaseModel as Base
-
+from .addons.base_selector import BaseSelector
+from .addons.definition_selector import DefinitionSelector
+from .addons.export_word_converter import ExportWordConverter
+from .addons.exporter import Exporter
+from .addons.key_selector import KeySelector
+from .addons.word_linker import WordLinker
+from .addons.word_selector import WordSelector
 from .author import BaseAuthor as Author
 from .definition import BaseDefinition as Definition
 from .event import BaseEvent as Event
 from .key import BaseKey as Key
+from .service.base import BaseModel as Base
 from .setting import BaseSetting as Setting
 from .syllable import BaseSyllable as Syllable
 from .type import BaseType as Type
 from .word import BaseWord as Word
 from .word_spell import BaseWordSpell as WordSpell
-
-from .addons.base_selector import BaseSelector
-from .addons.definition_selector import DefinitionSelector
-from .addons.key_selector import KeySelector
-from .addons.word_selector import WordSelector
-
-from .addons.export_word_converter import ExportWordConverter
-from .addons.exporter import Exporter
-from .addons.word_linker import WordLinker
