+++ loglan_core/service/annotated_types.py
@@ -1,0 +1,49 @@
+# pylint: disable=invalid-name
+"""
+This module provides custom type annotations for string lengths.
+"""
+
+from typing_extensions import Annotated
+
+
+str_008 = Annotated[str, 8]
+"""
+A custom type annotation that annotates a string with a metadata value 
+of 8. It can be used to associate a string with the number 8 in certain 
+contexts, such as specifying the length for a string field in a database.
+"""
+
+str_016 = Annotated[str, 16]
+"""
+A custom type annotation that annotates a string with a metadata value 
+of 16. It can be used to associate a string with the number 16 in certain 
+contexts, such as specifying the length for a string field in a database.
+"""
+
+str_032 = Annotated[str, 32]
+"""
+A custom type annotation that annotates a string with a metadata value 
+of 32. It can be used to associate a string with the number 32 in certain 
+contexts, such as specifying the length for a string field in a database.
+"""
+
+str_064 = Annotated[str, 64]
+"""
+A custom type annotation that annotates a string with a metadata value 
+of 64. It can be used to associate a string with the number 64 in certain 
+contexts, such as specifying the length for a string field in a database.
+"""
+
+str_128 = Annotated[str, 128]
+"""
+A custom type annotation that annotates a string with a metadata value 
+of 128. It can be used to associate a string with the number 128 in certain 
+contexts, such as specifying the length for a string field in a database.
+"""
+
+str_255 = Annotated[str, 255]
+"""
+A custom type annotation that annotates a string with a metadata value 
+of 255. It can be used to associate a string with the number 255 in certain 
+contexts, such as specifying the length for a string field in a database.
+"""
