+++ loglan_core/service/table_names.py
